# slacker-tracker/frontend/public/api-docs/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Slacker Tracker API 1.0",
    "description": "This is the api doc for slacker tracker",
    "contact": {
      "email": "mageguinsoo@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    }
  ],
...
  "paths": {
...
    "/friendList/": {
      "get": {
        "tags": [
          "friendList"
        ],
        "summary": "Get user's friendList",
        "description": "Get user's friendList",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "required": true,
            "description": "User's unique id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "_id": {
                        "type": "string",
                        "description": "The user ID."
                      },
                      "friendName": {
                        "type": "string",
                        "description": "The friend's name."
                      },
                      "email": {
                        "type": "string",
                        "description": "The friend's email."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "user not found"
          },
          "422": {
            "description": "id validation error"
          }
        }
      }
    },
...
    "/user/": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user's info",
        "description": "Get user's information",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "required": true,
            "description": "User's unique id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "_id": {
                        "type": "string",
                        "description": "The user ID."
                      },
                      "avatar": {
                        "type": "string",
                        "description": "The url for the user's avatar"
                      },
                      "email": {
                        "type": "string",
                        "description": "The user's email."
                      },
                      "bio": {
                        "type": "string",
                        "description": "The user's bio"
                      },
                      "name": {
                        "type": "string",
                        "description": "The user's name"
                      },
                      "slackerScore": {
                        "type": "number",
                        "description": "The user's slacker score"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "user not found"
          },
          "422": {
            "description": "id validation error"
          }
        }
      }
    },
...
    "/generateSummaryEmail": {
      "post": {
        "tags": [
          "summaryEmail"
        ],
        "summary": "Send summary email",
        "description": "Generate the weekly summary email and send to the specific email address",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "required": true,
            "description": "User's unique id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sendEmail",
            "in": "query",
            "required": true,
            "description": "The target email address",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "This account is registerd in this week"
          }
        }
      }
    }
  },
...
}